@@ -2,12 +2,7 @@ package com.github.tvbox.osc.bean;
 
 import java.util.List;
 
-/**
- * @Author : Liu XiaoRan
- * @Email : 592923276@qq.com
- * @Date : on 2023/8/15 10:58.
- * @Description :
- */
+
 public class VideoFolder {
     public VideoFolder(String name, List<VideoInfo> videoList) {
         this.name = name;
