@@ -1,12 +1,6 @@
 package com.github.tvbox.osc.constant;
 
-/**
- * @Author : Liu XiaoRan
- * @Email : 592923276@qq.com
- * @Date : on 2023/8/16 09:53.
- * @Description :
- */
-public class Constants {
+public class IntentKey {
 
     /**
      * 画中画/后台服务
@@ -16,4 +10,5 @@ public class Constants {
     public static final int BROADCAST_ACTION_PLAYPAUSE = 1;
     public static final int BROADCAST_ACTION_NEXT = 2;
     public static final int BROADCAST_ACTION_CLOSE = 3;
+    public static final String CACHE_CONFIG_CHANGED = "use_cache";
 }
