@@ -5,7 +5,6 @@ import android.content.Context;
 import android.content.Intent;
 
 import com.github.tvbox.osc.event.ServerEvent;
-import com.github.tvbox.osc.ui.activity.SearchActivity;
 import com.github.tvbox.osc.util.AppManager;
 
 import org.greenrobot.eventbus.EventBus;
@@ -20,16 +19,6 @@ public class SearchReceiver extends BroadcastReceiver {
 
     @Override
     public void onReceive(Context context, Intent intent) {
-        if (action.equals(intent.getAction()) && intent.getExtras() != null) {
-            if (AppManager.getInstance().getActivity(SearchActivity.class) != null) {
-                AppManager.getInstance().backActivity(SearchActivity.class);
-                EventBus.getDefault().post(new ServerEvent(ServerEvent.SERVER_SEARCH, intent.getExtras().getString("title")));
-            } else {
-                Intent newIntent = new Intent(context, SearchActivity.class);
-                newIntent.putExtra("title", intent.getExtras().getString("title"));
-                newIntent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TOP);
-                context.startActivity(newIntent);
-            }
-        }
+
     }
 }
