@@ -23,12 +23,7 @@ import com.squareup.picasso.Picasso;
 
 import me.jessyan.autosize.utils.AutoSizeUtils;
 
-/**
- * @Author : Liu XiaoRan
- * @Email : 592923276@qq.com
- * @Date : on 2023/8/17 09:28.
- * @Description :
- */
+
 public class VideoDetailDialog extends BottomPopupView {
 
 
