@@ -3,7 +3,6 @@ package com.github.tvbox.osc.util;
 import com.github.tvbox.osc.api.ApiConfig;
 import com.github.tvbox.osc.bean.SourceBean;
 import com.github.tvbox.osc.ui.activity.FastSearchActivity;
-import com.github.tvbox.osc.ui.activity.SearchActivity;
 import com.orhanobut.hawk.Hawk;
 
 import java.util.ArrayList;
@@ -41,8 +40,7 @@ public class SearchHelper {
             if (mCheckSourcesForApi == null) mCheckSourcesForApi = new HashMap<>();
             mCheckSourcesForApi.put(api, mCheckSources);
         }
-        SearchActivity.setCheckedSourcesForSearch(mCheckSources);
-        FastSearchActivity.setCheckedSourcesForSearch(mCheckSources);
+        FastSearchActivity.Companion.setCheckedSourcesForSearch(mCheckSources);
         Hawk.put(HawkConfig.SOURCES_FOR_SEARCH, mCheckSourcesForApi);
     }
 
